Revert "#76 Don't provision additional capacity until node will be online"
This reverts commit 545a25fc
